requirements.txt
# =============================================================================
# PYTHON DEPENDENCIES FOR DATA PLATFORM
# =============================================================================
# Airflow 2.8.1 + Spark 3.5.0 + MinIO Integration
# =============================================================================

# -----------------------------------------------------------------------------
# APACHE SPARK & PYSPARK
# -----------------------------------------------------------------------------
pyspark==3.5.0
py4j==0.10.9.7

# -----------------------------------------------------------------------------
# AIRFLOW PROVIDERS
# -----------------------------------------------------------------------------
apache-airflow-providers-amazon>=8.0.0
apache-airflow-providers-postgres>=5.0.0
apache-airflow-providers-celery>=3.0.0
apache-airflow-providers-redis>=3.0.0
apache-airflow-providers-http>=4.0.0

# -----------------------------------------------------------------------------
# OBJECT STORAGE (MinIO / S3)
# -----------------------------------------------------------------------------
boto3==1.34.10
botocore==1.34.10
s3fs==2024.2.0
minio==7.2.3

# -----------------------------------------------------------------------------
# DATA PROCESSING & ANALYTICS
# -----------------------------------------------------------------------------
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2
fastparquet==2023.10.1
openpyxl==3.1.2
xlrd==2.0.1

# -----------------------------------------------------------------------------
# DATABASE CONNECTORS
# -----------------------------------------------------------------------------
psycopg2-binary==2.9.9
SQLAlchemy>=1.4,<2.0
pymongo==4.6.1

# -----------------------------------------------------------------------------
# DATA VALIDATION & QUALITY
# -----------------------------------------------------------------------------
pydantic>=2.0,<3.0

# -----------------------------------------------------------------------------
# HTTP & API
# -----------------------------------------------------------------------------
requests==2.31.0
httpx==0.26.0

# -----------------------------------------------------------------------------
# UTILITIES
# -----------------------------------------------------------------------------
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3.post1

# -----------------------------------------------------------------------------
# DATA SCIENCE & ML (Optional)
# -----------------------------------------------------------------------------
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.1

# -----------------------------------------------------------------------------
# DEVELOPMENT & DEBUGGING
# -----------------------------------------------------------------------------
ipython==8.19.0
jupyter==1.0.0

# -----------------------------------------------------------------------------
# PERFORMANCE & OPTIMIZATION
# -----------------------------------------------------------------------------
joblib==1.3.2
tqdm==4.66.1

# -----------------------------------------------------------------------------
# FILE FORMAT SUPPORT
# -----------------------------------------------------------------------------
avro-python3==1.10.2
xmltodict==0.13.0

# -----------------------------------------------------------------------------
# SECURITY
# -----------------------------------------------------------------------------
cryptography==41.0.7